custom UI hue + accent, and the Torii gate glyph as the settings icon
- custom UI hue: ported CustomUiHueHelper + FormHuePicker + APIUser.HasSupported (donator gate),
  applied centrally in FullscreenOverlay (covers chat/login/profile/rankings/etc.) via
  BindFullScheme with live re-tint. settable in Settings -> Torii -> Interface (hue + separate
  accent hue + enable toggles). hue settings are float to match the helper/picker.
- settings section icon is now the Torii gate glyph in the red->yellow dashboard gradient
  (was a placeholder bolt).

+++ osu.Game/Configuration/OsuConfigManager.cs
@@ -189,13 +189,13 @@ namespace osu.Game.Configuration
             SetDefault(OsuSetting.CosmeticsHidden, false);
             SetDefault(OsuSetting.ToriiPointsBalance, 0);
             SetDefault(OsuSetting.ToriiPointsFeedCursor, 0);
-            SetDefault(OsuSetting.CustomUIHue, osu.Game.Overlays.OverlayColourScheme.Purple.GetHue(), 0, 359, 1);
+            SetDefault(OsuSetting.CustomUIHue, (float)osu.Game.Overlays.OverlayColourScheme.Purple.GetHue(), 0f, 359f, 1f);
             SetDefault(OsuSetting.CustomUIHueEnabled, false);
             SetDefault(OsuSetting.CustomUIHueApplyToMenu, false);
             SetDefault(OsuSetting.CustomUIHueApplyToOverlays, true);
             SetDefault(OsuSetting.CustomUIHueApplyToSettingsPanel, true);
             SetDefault(OsuSetting.CustomUIAccentEnabled, false);
-            SetDefault(OsuSetting.CustomUIAccentHue, osu.Game.Overlays.OverlayColourScheme.Purple.GetHue(), 0, 359, 1);
+            SetDefault(OsuSetting.CustomUIAccentHue, (float)osu.Game.Overlays.OverlayColourScheme.Purple.GetHue(), 0f, 359f, 1f);
             SetDefault(OsuSetting.CustomUIAccentUnlocked, false);
             SetDefault(OsuSetting.MenuCursorStyle, osu.Game.Graphics.Cursor.MenuCursorStyle.LazerDefault);
             SetDefault(OsuSetting.UseGameplayCursorInMenus, false);
